requirements.txt
flask==3.0.0
mcstatus==11.1.0
Pillow==11.3.0
requests==2.32.4